Add batch link capture components
